requirements.txt
llama-index
llama-hub
transformers
torch
pytesseract
Pillow
pdf2image
arize-phoenix[experimental]
docx2txt
xlrd
svglib
streamlit
watchdog
streamlit-authenticator